adding readme instructions and minor changes
