+++ app/pages/1_Scenario_comparison.py
@@ -2,10 +2,10 @@ import streamlit as st
 st.set_page_config(
     layout="wide"
 )
-import app.pages.utils.tools as tools
 import pathlib
 import plotly.express as px
 import app.pages.utils.tools as tools
+import app.pages.utils.scenario_comparision_prerun as helper
 
 
 def get_stat_unit(param):
@@ -22,8 +22,8 @@ def main():
         "Select your metric",
         params,
     )
-    df = tools.get_df_for_parameter(
-        network_map, option, tools.add_values_for_statistics, tools.get_stats_col_names
+    df = helper.get_df_for_parameter(
+        network_map, option, helper.add_values_for_statistics, helper.get_stats_col_names
     )
     _,plot_col,_=st.columns([1,80,1])
     with plot_col:
@@ -40,11 +40,11 @@ def main():
     )
 
     if option == "CO2 emissions":
-        co2_df = tools.get_df_for_parameter(
+        co2_df = helper.get_df_for_parameter(
             network_map,
             "co2_emissions",
-            tools.add_values_for_co2,
-            tools.get_co2_col_names,
+            helper.add_values_for_co2,
+            helper.get_co2_col_names,
         )
 
         _,plot_col,_=st.columns([1,60,1])
@@ -57,8 +57,8 @@ def main():
             st.plotly_chart(fig, use_container_width=True)
 
     elif option == "optimal generator capacity":
-        gen_df = tools.get_df_for_parameter(
-            network_map, None, tools.add_values_for_generators, tools.get_gen_col_names
+        gen_df = helper.get_df_for_parameter(
+            network_map, None, helper.add_values_for_generators, helper.get_gen_col_names
         )
 
         _,plot_col,_=st.columns([1,80,1])
